Fixed insidious issue in partition
After days of trying to figure out why loop was misbehaving, I finally
realized the issue was actually in the implementation of partition.
By calling push! directly on an empty list created with () rather than
(list), I was actually modifying the AST itself.

The issue is fixed, and I've changed the mutating functions to all have
dollar-signs and exclamation marks to denote that they're fucking
dangerous and not to be trifled with.